requirements.txt
transformers>=4.30.0
torch>=2.0.0
pdf2image>=1.16.3
pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
camelot-py[base]>=0.11.0
tabula-py>=2.8.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
openpyxl>=3.1.0
pytesseract>=0.3.0
openpyxl-image-loader>=1.0.0
xlwings>=0.28.0
spire-pdf==11.9.2
streamlit==1.47.1
